refactor: move types, change orderQuantityInKG to Quantites, change type to an object
