+++ packages/certificate-summary/src/index.ts
@@ -7,6 +7,7 @@ import {
   PartyEmail,
 } from '@s1seven/schema-tools-extract-emails';
 import { CoASchema, ECoCSchema, EN10168Schema, SupportedSchemas } from '@s1seven/schema-tools-types';
+import { EcocData, Quantity } from '@s1seven/schema-tools-types/src';
 import { castCertificate, loadExternalFile } from '@s1seven/schema-tools-utils';
 
 export interface CertificateSummary {
@@ -18,23 +19,8 @@ export interface CertificateSummary {
   purchaseDeliveryPosition?: string;
   purchaseOrderNumber?: string;
   purchaseOrderPosition?: string;
-  orderQuantityInKG?: string;
+  quantity?: Quantity[] | Quantity[][];
 }
-
-interface EcocData {
-  Data: {
-    ObjectOfDeclaration: ObjectOfDeclaration[];
-  };
-}
-
-type ObjectOfDeclaration = {
-  Quantities: Quantities[];
-};
-
-type Quantities = {
-  Amount: number;
-  Unit: string;
-};
 
 function getFirstPartyName(parties: PartyEmail[]): string {
   return parties?.length ? parties[0]?.name : '';
@@ -56,9 +42,10 @@ function extractSummaryFromEN10168(certificate: EN10168Schema): CertificateSumma
   const purchaseDeliveryNumber = Object.prototype.hasOwnProperty.call(CommercialTransaction, 'A98')
     ? CommercialTransaction['A98'].toString()
     : '';
-  const orderQuantityInKG = Object.prototype.hasOwnProperty.call(ProductDescription, 'B13')
-    ? ProductDescription['B13']['Value'].toString()
-    : '';
+
+  const quantity: Quantity[] = Object.prototype.hasOwnProperty.call(ProductDescription, 'B13')
+    ? [{ Value: ProductDescription['B13']['Value'], Unit: ProductDescription['B13']['Unit'] }]
+    : [];
 
   return {
     certificateIdentifier: certificate.Certificate.CommercialTransaction.A03,
@@ -69,7 +56,7 @@ function extractSummaryFromEN10168(certificate: EN10168Schema): CertificateSumma
     purchaseDeliveryPosition: undefined,
     purchaseOrderNumber,
     purchaseOrderPosition,
-    orderQuantityInKG,
+    quantity,
   };
 }
 
@@ -86,16 +73,16 @@ function extractSummaryFromECoC(certificate: ECoCSchema): CertificateSummary | n
   const purchaseDeliveryNumber = StandardReferences?.find((ref) => ref?.name === 'DeliveryNote')?.Value;
 
   const ecocData = certificate.EcocData as EcocData;
-  const orderQuantityInKG = (
-    ecocData?.Data?.ObjectOfDeclaration?.reduce((acc, objOfDeclarartion) => {
-      return (
-        acc +
-        objOfDeclarartion?.Quantities?.reduce((acc, quantitiesObj) => {
-          return acc + (quantitiesObj?.Amount || 0);
-        }, 0)
-      );
-    }, 0) || ''
-  ).toString();
+  const quantity =
+    ecocData?.Data?.ObjectOfDeclaration?.reduce<Quantity[][]>((acc, ObjectOfDeclaration) => {
+      return (acc = [
+        ...acc,
+        ObjectOfDeclaration.Quantities.map<Quantity>((quantityObj) => {
+          const { Amount: Value, Unit } = quantityObj;
+          return { Value, Unit };
+        }),
+      ]);
+    }, []) || [];
 
   return {
     certificateIdentifier: certificate.Id,
@@ -106,7 +93,7 @@ function extractSummaryFromECoC(certificate: ECoCSchema): CertificateSummary | n
     purchaseDeliveryPosition: undefined,
     purchaseOrderNumber,
     purchaseOrderPosition,
-    orderQuantityInKG,
+    quantity,
   };
 }
 
@@ -126,7 +113,7 @@ function extractSummaryFromCoA(certificate: CoASchema): CertificateSummary | nul
   const purchaseOrderPosition = Order?.Position;
   const purchaseDeliveryNumber = Delivery?.Number || Delivery?.Id;
   const purchaseDeliveryPosition = Delivery?.Position;
-  const orderQuantityInKG = Delivery?.Quantity.toString();
+  const quantity = [{ Unit: Delivery?.QuantityUnit, Value: Delivery?.Quantity }];
   return {
     certificateIdentifier: certificate.Certificate.Id,
     sellerName: getFirstPartyName(senders),
@@ -136,7 +123,7 @@ function extractSummaryFromCoA(certificate: CoASchema): CertificateSummary | nul
     purchaseDeliveryPosition,
     purchaseOrderNumber,
     purchaseOrderPosition,
-    orderQuantityInKG,
+    quantity,
   };
 }
